Move icons from images to icons folder
